python/requirements.txt
py_ecc
pycryptodome
eth-abi
web3